Configuration options for embedded use
Allow the user to explicitly configure the producer and the metric registry in
code. This allows Maxwell to be used as a library within another application,
and allows extra flexibility for customization of Maxwell's behavior.

The producer is the main piece that people will want to customize if they
want to send metrics to a different destination, use a different data format,
use a custom partitioning strategy, handle failures differently, etc. (I think
there is more we could do to make this easy, by allowing extension or reuse
of MaxwellKafkaProducer, but it's not done yet.)

The static MetricRegistry and HealthCheckRegistry were also converted to be
non-static and part of the configuration, because I want to share these with
the container application.

Also, move the Maxwell application's logger initialization into the main method
instead of a static init block, so that it doesn't interfere with a host
application's logging configuration.